progress snapshot
